Cleanup scene_tests/test_load.py
